move `type_op` into `rustc`
