+++ src/librustc/traits/query/type_op/outlives.rs
@@ -8,41 +8,41 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use rustc::infer::{InferCtxt, InferResult};
-use rustc::traits::ObligationCause;
-use rustc::ty::{ParamEnv, Ty, TyCtxt};
+use infer::{InferCtxt, InferResult};
+use traits::query::dropck_outlives::trivial_dropck_outlives;
+use traits::ObligationCause;
+use ty::subst::Kind;
+use ty::{ParamEnv, Ty, TyCtxt};
 
 #[derive(Debug)]
-crate struct Subtype<'tcx> {
+pub struct DropckOutlives<'tcx> {
     param_env: ParamEnv<'tcx>,
-    sub: Ty<'tcx>,
-    sup: Ty<'tcx>,
+    dropped_ty: Ty<'tcx>,
 }
 
-impl<'tcx> Subtype<'tcx> {
-    crate fn new(param_env: ParamEnv<'tcx>, sub: Ty<'tcx>, sup: Ty<'tcx>) -> Self {
-        Self {
+impl<'tcx> DropckOutlives<'tcx> {
+    pub fn new(param_env: ParamEnv<'tcx>, dropped_ty: Ty<'tcx>) -> Self {
+        DropckOutlives {
             param_env,
-            sub,
-            sup,
+            dropped_ty,
         }
     }
 }
 
-impl<'gcx, 'tcx> super::TypeOp<'gcx, 'tcx> for Subtype<'tcx> {
-    type Output = ();
+impl<'gcx, 'tcx> super::TypeOp<'gcx, 'tcx> for DropckOutlives<'tcx> {
+    type Output = Vec<Kind<'tcx>>;
 
-    fn trivial_noop(self, _tcx: TyCtxt<'_, 'gcx, 'tcx>) -> Result<Self::Output, Self> {
-        if self.sub == self.sup {
-            Ok(())
+    fn trivial_noop(self, tcx: TyCtxt<'_, 'gcx, 'tcx>) -> Result<Self::Output, Self> {
+        if trivial_dropck_outlives(tcx, self.dropped_ty) {
+            Ok(vec![])
         } else {
             Err(self)
         }
     }
 
     fn perform(self, infcx: &InferCtxt<'_, 'gcx, 'tcx>) -> InferResult<'tcx, Self::Output> {
-        infcx
+        Ok(infcx
             .at(&ObligationCause::dummy(), self.param_env)
-            .sup(self.sup, self.sub)
+            .dropck_outlives(self.dropped_ty))
     }
 }
